[refactor, feat] modularize migrationprocess, add unprogrammed account details form
